refactor : LottoNumber 와 LottoCache 를 통합하여 LottoNumber 에서 of로 새로운 객체를 생성하는 것에 대해서 보장가능하게 변경함
