+++ README.md
@@ -180,7 +180,7 @@
       - [x] : 넘었을 경우 다시 입력 유도
 - [x] : 수동으로 구매할 로또 번호 입력
   - [x] : 수동 로또 수만큼 반복입력한다
-    - [ ] : 중복 입력, 1 ~ 45 범위 넘겨서 입력 시 다시 입력 유도
+      - [x] : 중복 입력, 1 ~ 45 범위 넘겨서 입력 시 다시 입력 유도
       - 고민 : 입력 한번 할 때 마다 제시 or 모아서 넘기기
       - [x] : 모아서 넘기기
 - [x] : 로또 발행 결과 수동, 자동 갯수 카운트
@@ -203,3 +203,10 @@
 #### 추가
 - `BuyLotto` : 로또 구매(자동 + 수동) 에 대한 관리
 - `manualLotto` : 수동구매하는 로또와 그 갯수 관리
+
+## 1차 피드백 후 리팩터링
+
+- [ ] : 로또를 생성하는 부분 인터페이스로 분리해 보는 연습
+- [x] : of 메서드를 활용하면 LottoNumber 객체가 재사용되는 것을 보장하지 못함
+    - `LottoNumber` 객체가 캐싱을 통해 재사용하는 것을 강제할 수 있도록 리팩터링
+    - `LottoNumber`와 `LottoNumberCache`를 통합
