+++ src/main/java/lotto/domain/model/WinningLotto.java
@@ -5,19 +5,19 @@ import lotto.domain.Rank;
 public record WinningLotto(LottoNumber bonusNumber, Lotto winNumbers) {
     
     public WinningLotto(int bonusNumber, int... winNumbers) {
-        this(LottoNumberCache.getLottoNumber(bonusNumber), new Lotto(winNumbers));
+        this(LottoNumber.of(bonusNumber), new Lotto(winNumbers));
     }
     
     public WinningLotto(String bonusNumber, int... winNumbers) {
-        this(LottoNumberCache.getLottoNumber(bonusNumber), new Lotto(winNumbers));
+        this(LottoNumber.of(bonusNumber), new Lotto(winNumbers));
     }
     
     public WinningLotto(String bonusNumber, String winNumbers) {
-        this(LottoNumberCache.getLottoNumber(bonusNumber), new Lotto(winNumbers));
+        this(LottoNumber.of(bonusNumber), new Lotto(winNumbers));
     }
     
     public WinningLotto(int bonusNumber, String winNumbers) {
-        this(LottoNumberCache.getLottoNumber(bonusNumber), new Lotto(winNumbers));
+        this(LottoNumber.of(bonusNumber), new Lotto(winNumbers));
     }
     
     public WinningLotto {
